fix(audio): add [PreserveSig] to the ducking COM interop; fix duration overflow
Review H8 + M3:
- Every method on the four DuckingOptOut [ComImport] interfaces returns an
  int HRESULT that we check via Marshal.ThrowExceptionForHR — that contract
  requires [PreserveSig]. Without it the marshaler rewrites each signature
  (hidden HRESULT + retval argument), so the checks read garbage and x86
  would be a stack imbalance. Worked on x64 by accident.
- Recorder duration math now uses long: trimmed.Length * 1000 overflowed
  int at ~44.7 s, persisting a negative DurationMs for long script takes.

+++ src/AdaVoice.Audio/Recording/Recorder.cs
@@ -76,7 +76,7 @@ public sealed class Recorder
 
         var referenceRms = _options.ReferenceRms ?? RampGain.DbToLinear(_options.ReferenceDbfs);
         var gainDb = LoudnessMatch.ComputeGainDb(trimmed, referenceRms, _options.PeakCeilingDbfs);
-        var durationMs = trimmed.Length * 1000 / AudioFormats.SampleRate;
+        var durationMs = (int)(trimmed.Length * 1000L / AudioFormats.SampleRate); // long math: int overflows at ~44.7 s
         var peakDbfs = 20 * Math.Log10(Math.Max(Loudness.Peak(trimmed), 1e-9));
 
         return new RecordingResult(trimmed, gainDb, durationMs, peakDbfs);
